requirements.txt
rumps
TreeStructConfig
googletrans==4.0.0rc1

pyobjc-framework-Cocoa
pyobjc-framework-CoreServices
pyobjc-framework-WebKit

sentry-sdk

# for build
wheel
py2app
click
